requirements.txt
Django>=3.2.21,<4.0.0
Pillow==10.0.*
parsy==1.4.0
django-google-tag-manager==0.0.5
markdown==3.4.*
hypothesis>=6.84.2,<6.85.0
requests==2.31.*
deal==4.24.*
django-stubs>=4.2.4,<4.3.0
django-stubs-ext>=4.2.2,<4.3.0
typing-extensions>=4.4.0,<4.5.0
django-geojson==4.0.*
django-mptt==0.14.*
